+++ src/odds.rs
@@ -1,6 +1,4 @@
-use libc::c_int;
-
-pub const WAKE_PERCENT: c_int = 45;
-pub const FLIT_PERCENT: c_int = 33;
-pub const PARTY_WAKE_PERCENT: c_int = 75;
-pub const STEALTH_FACTOR: c_int = 3;
+pub const WAKE_PERCENT: usize = 45;
+pub const FLIT_PERCENT: usize = 33;
+pub const PARTY_WAKE_PERCENT: usize = 75;
+pub const STEALTH_FACTOR: usize = 3;
